Update repositories.py
Hi Helge. We have changed our minitwit CI/CD pipeline and therefore also have a new URL to the app/api. The old one was just the app and api running in a container on azure. The problem we had was related to the fact that we had included the SQLite DB in the container, instead of setting up a server and DB remotely. This new app/api is deployed with azure DevOps pipelines is setup to connect to a remote DB, however we have been unable to salvage the data from the SQLite DB running in the old container. We realize that loosing all the old data is a problem, and we have tried to retrieve it but without success.